Move all extension methods next to the appropriate StreamTransformer
