+++ lib/src/subjects/behavior_subject.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:rxdart/src/extensions.dart';
 import 'package:rxdart/src/observables/observable.dart';
 import 'package:rxdart/src/observables/value_observable.dart';
 import 'package:rxdart/src/subjects/subject.dart';
