+++ lib/src/transformers/backpressure/buffer.dart
@@ -93,3 +93,67 @@ class BufferTestStreamTransformer<T>
     if (test == null) throw ArgumentError.notNull('test');
   }
 }
+
+extension BufferExtensions<T> on Stream<T> {
+  /// Creates a Stream where each item is a [List] containing the items
+  /// from the source sequence.
+  ///
+  /// This [List] is emitted every time [window] emits an event.
+  ///
+  /// ### Example
+  ///
+  ///     Stream.periodic(Duration(milliseconds: 100), (i) => i)
+  ///       .buffer(Stream.periodic(Duration(milliseconds: 160), (i) => i))
+  ///       .listen(print); // prints [0, 1] [2, 3] [4, 5] ...
+  Stream<List<T>> buffer(Stream window) =>
+      transform(BufferStreamTransformer((_) => window));
+
+  /// Buffers a number of values from the source Stream by [count] then
+  /// emits the buffer and clears it, and starts a new buffer each
+  /// [startBufferEvery] values. If [startBufferEvery] is not provided,
+  /// then new buffers are started immediately at the start of the source
+  /// and when each buffer closes and is emitted.
+  ///
+  /// ### Example
+  /// [count] is the maximum size of the buffer emitted
+  ///
+  ///     RangeStream(1, 4)
+  ///       .bufferCount(2)
+  ///       .listen(print); // prints [1, 2], [3, 4] done!
+  ///
+  /// ### Example
+  /// if [startBufferEvery] is 2, then a new buffer will be started
+  /// on every other value from the source. A new buffer is started at the
+  /// beginning of the source by default.
+  ///
+  ///     RangeStream(1, 5)
+  ///       .bufferCount(3, 2)
+  ///       .listen(print); // prints [1, 2, 3], [3, 4, 5], [5] done!
+  Stream<List<T>> bufferCount(int count, [int startBufferEvery = 0]) =>
+      transform(BufferCountStreamTransformer<T>(count, startBufferEvery));
+
+  /// Creates a Stream where each item is a [List] containing the items
+  /// from the source sequence, batched whenever test passes.
+  ///
+  /// ### Example
+  ///
+  ///     Stream.periodic(Duration(milliseconds: 100), (int i) => i)
+  ///       .bufferTest((i) => i % 2 == 0)
+  ///       .listen(print); // prints [0], [1, 2] [3, 4] [5, 6] ...
+  Stream<List<T>> bufferTest(bool onTestHandler(T event)) =>
+      transform(BufferTestStreamTransformer<T>(onTestHandler));
+
+  /// Creates a Stream where each item is a [List] containing the items
+  /// from the source sequence, sampled on a time frame with [duration].
+  ///
+  /// ### Example
+  ///
+  ///     Stream.periodic(Duration(milliseconds: 100), (int i) => i)
+  ///       .bufferTime(Duration(milliseconds: 220))
+  ///       .listen(print); // prints [0, 1] [2, 3] [4, 5] ...
+  Stream<List<T>> bufferTime(Duration duration) {
+    if (duration == null) throw ArgumentError.notNull('duration');
+
+    return buffer(Stream<void>.periodic(duration));
+  }
+}
