+++ lib/src/transformers/backpressure/pairwise.dart
@@ -18,3 +18,14 @@ class PairwiseStreamTransformer<T>
             closeWindowWhen: (Iterable<T> queue) => queue.length == 2,
             dispatchOnClose: false);
 }
+
+extension PairwiseExtension<T> on Stream<T> {
+  /// Emits the n-th and n-1th events as a pair..
+  ///
+  /// ### Example
+  ///
+  ///     RangeStream(1, 4)
+  ///       .pairwise()
+  ///       .listen(print); // prints [1, 2], [2, 3], [3, 4]
+  Stream<Iterable<T>> pairwise() => transform(PairwiseStreamTransformer());
+}
