+++ lib/src/transformers/backpressure/throttle.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:rxdart/src/streams/timer.dart';
 import 'package:rxdart/src/transformers/backpressure/backpressure.dart';
 
 /// A [StreamTransformer] that emits only the first item emitted by the source
@@ -23,4 +24,35 @@ class ThrottleStreamTransformer<T> extends BackpressureStreamTransformer<T, T> {
             onWindowEnd: trailing ? (Iterable<T> queue) => queue.last : null) {
     assert(window != null, 'window stream factory cannot be null');
   }
+}
+
+extension ThrottleExtensions<T> on Stream<T> {
+  /// Emits only the first item emitted by the source [Stream] while [window] is
+  /// open.
+  ///
+  /// if [trailing] is true, then the last item is emitted instead
+  ///
+  /// You can use the value of the last throttled event to determine the length
+  /// of the next [window].
+  ///
+  /// ### Example
+  ///
+  ///     Stream.fromIterable([1, 2, 3])
+  ///       .throttle((_) => TimerStream(true, Duration(seconds: 1)))
+  Stream<T> throttle(Stream window(T event), {bool trailing = false}) =>
+      transform(ThrottleStreamTransformer<T>(window, trailing: trailing));
+
+  /// Emits only the first item emitted by the source [Stream] within a time
+  /// span of [duration].
+  ///
+  /// if [trailing] is true, then the last item is emitted instead
+  ///
+  /// ### Example
+  ///
+  ///     Stream.fromIterable([1, 2, 3])
+  ///       .throttleTime(Duration(seconds: 1))
+  Stream<T> throttleTime(Duration duration, {bool trailing = false}) =>
+      transform(ThrottleStreamTransformer<T>(
+          (_) => TimerStream<bool>(true, duration),
+          trailing: trailing));
 }
