+++ lib/src/transformers/exhaust_map.dart
@@ -82,3 +82,21 @@ class ExhaustMapStreamTransformer<T, S> extends StreamTransformerBase<T, S> {
     });
   }
 }
+
+extension ExhaustMapExtension<T> on Stream<T> {
+  /// Converts items from the source stream into a Stream using a given
+  /// mapper. It ignores all items from the source stream until the new stream
+  /// completes.
+  ///
+  /// Useful when you have a noisy source Stream and only want to respond once
+  /// the previous async operation is finished.
+  ///
+  /// ### Example
+  ///
+  ///     RangeStream(0, 2).interval(Duration(milliseconds: 50))
+  ///       .exhaustMap((i) =>
+  ///         TimerStream(i, Duration(milliseconds: 75)))
+  ///       .listen(print); // prints 0, 2
+  Stream<S> exhaustMap<S>(Stream<S> mapper(T value)) =>
+      transform(ExhaustMapStreamTransformer<T, S>(mapper));
+}
