+++ lib/src/transformers/map_to.dart
@@ -38,3 +38,15 @@ class MapToStreamTransformer<S, T> extends StreamTransformerBase<S, T> {
         return controller.stream.listen(null);
       });
 }
+
+extension MapToExtension<T> on Stream<T> {
+  /// Emits the given constant value on the output Stream every time the source
+  /// Stream emits a value.
+  ///
+  /// ### Example
+  ///
+  ///     Stream.fromIterable([1, 2, 3, 4])
+  ///       .mapTo(true)
+  ///       .listen(print); // prints true, true, true, true
+  Stream<S> mapTo<S>(S value) => transform(MapToStreamTransformer<T, S>(value));
+}
