+++ lib/src/transformers/take_until.dart
@@ -64,3 +64,19 @@ class TakeUntilStreamTransformer<T, S> extends StreamTransformerBase<T, T> {
     });
   }
 }
+
+extension TakeUntilExtension<T> on Stream<T> {
+  /// Returns the values from the source Stream sequence until the other Stream
+  /// sequence produces a value.
+  ///
+  /// ### Example
+  ///
+  ///     MergeStream([
+  ///         Stream.fromIterable([1]),
+  ///         TimerStream(2, Duration(minutes: 1))
+  ///       ])
+  ///       .takeUntil(TimerStream(3, Duration(seconds: 10)))
+  ///       .listen(print); // prints 1
+  Stream<T> takeUntil<S>(Stream<S> otherStream) =>
+      transform(TakeUntilStreamTransformer<T, S>(otherStream));
+}
