+++ lib/src/transformers/where_type.dart
@@ -56,3 +56,25 @@ class WhereTypeStreamTransformer<S, T> extends StreamTransformerBase<S, T> {
         return controller.stream.listen(null);
       });
 }
+
+extension WhereTypeExtension<T> on Stream<T> {
+  /// This transformer is a shorthand for [Stream.where] followed by
+  /// [Stream.cast].
+  ///
+  /// Events that do not match [T] are filtered out, the resulting [Stream] will
+  /// be of Type [T].
+  ///
+  /// ### Example
+  ///
+  ///     Stream.fromIterable([1, 'two', 3, 'four'])
+  ///       .whereType<int>()
+  ///       .listen(print); // prints 1, 3
+  ///
+  /// #### as opposed to:
+  ///
+  ///     Stream.fromIterable([1, 'two', 3, 'four'])
+  ///       .where((event) => event is int)
+  ///       .cast<int>()
+  ///       .listen(print); // prints 1, 3
+  Stream<S> whereType<S>() => transform(WhereTypeStreamTransformer<T, S>());
+}
